Renamed SidebarData to DashboardSidebarData
